+++ src/utils/index.ts
@@ -4,7 +4,12 @@ import { MFE, MFME } from './messages'
 import { PATA } from './constants'
 import { decryptMessage, encryptMessage, generatePassword } from './security'
 import { response } from './response'
-import { signAccessToken, verifyAccessToken } from './jwt.helper'
+import {
+  signAccessToken,
+  signRefreshToken,
+  verifyAccessToken,
+  verifyRefreshToken
+} from './jwt.helper'
 
 export {
   docs,
@@ -17,5 +22,7 @@ export {
   generatePassword,
   response,
   signAccessToken,
-  verifyAccessToken
+  signRefreshToken,
+  verifyAccessToken,
+  verifyRefreshToken
 }