+++ src/utils/jwt.helper.ts
@@ -1,4 +1,4 @@
-/* eslint-disable no-extra-parens */
+/* eslint-disable @typescript-eslint/ban-types, @typescript-eslint/no-explicit-any, no-extra-parens */
 import httpErrors from 'http-errors'
 import jwt from 'jsonwebtoken'
 import { Response, NextFunction } from 'express'
@@ -6,9 +6,7 @@ import { Request } from '../custom/express.request'
 
 const signAccessToken = (id: string): Promise<unknown> => {
   return new Promise((resolve, reject) => {
-    const payload = {
-      name: 'Yours truly'
-    }
+    const payload = {}
     const secret = process.env.ACCESS_TOKEN_SECRET as string
 
     const options = {
@@ -25,12 +23,33 @@ const signAccessToken = (id: string): Promise<unknown> => {
   })
 }
 
+const signRefreshToken = (id: string): Promise<unknown> => {
+  return new Promise((resolve, reject) => {
+    const payload = {}
+    const secret = process.env.REFRESH_TOKEN_SECRET as string
+
+    const options = {
+      audience : id,
+      expiresIn: '1d',
+      issuer   : process.env.RHOADES_FRONT_URL as string
+    }
+    jwt.sign(payload, secret, options, (error, token): void => {
+      if (error) {
+        console.error(error)
+        reject(new httpErrors.InternalServerError('Ups! Something went wrong'))
+      } else resolve(token)
+    })
+  })
+}
+
 const verifyAccessToken = (
-  req : Request,
-  res : Response,
+  req: Request,
+  res: Response,
   next: NextFunction
 ): void => {
-  const { headers: { authorization } } = req
+  const {
+    headers: { authorization }
+  } = req
   if (!authorization) next(new httpErrors.Unauthorized())
 
   const bearerToken = (authorization as string).split(' ')
@@ -50,4 +69,24 @@ const verifyAccessToken = (
   )
 }
 
-export { signAccessToken, verifyAccessToken }
+const verifyRefreshToken = (refreshToken: string): Promise<unknown> => {
+  return new Promise((resolve, reject) => {
+    jwt.verify(
+      refreshToken,
+      process.env.REFRESH_TOKEN_SECRET as string,
+      (error, payload) => {
+        if (error) reject(new httpErrors.Unauthorized())
+        const userId = (payload as any).aud
+
+        resolve(userId)
+      }
+    )
+  })
+}
+
+export {
+  signAccessToken,
+  signRefreshToken,
+  verifyAccessToken,
+  verifyRefreshToken
+}